change from require and :refer to use and :only where appropriate
You use `require` and `:refer` when you need access to the entire module, but also want access to a few commonly used methods in the current namespace. If you are only using a couple of functions, it is more idiomatic to simply `use` them.

As a rule, if you aren't also using the :as keyword with `require` because you need a short name for that module in the current namespace, then you shouldn't be using refer, you should be using `use` and `:only`.